Changed HP, Mana, and Endurance to use int16 instead of int32 to prevent clients crashing or incorrectly dying.

+++ zone/client.cpp
@@ -1250,7 +1250,7 @@ void Client::UpdateAdmin(bool iFromDB) {
 	UpdateWho();
 }
 
-const int32& Client::SetMana(int32 amount) {
+const int16& Client::SetMana(int16 amount) {
 	bool update = false;
 	if (amount < 0)
 		amount = 0;
@@ -2787,7 +2787,7 @@ void Client::SendRules(Client* client)
 	safe_delete_array(rules);
 }
 
-void Client::SetEndurance(int32 newEnd)
+void Client::SetEndurance(int16 newEnd)
 {
 	/*Endurance can't be less than 0 or greater than max*/
 	if(newEnd < 0)
