+++ zone/npc.h
@@ -202,7 +202,7 @@ public:
 	inline void	SetPlatinum(uint32 amt)		{ platinum = amt; }
 
 
-	virtual int32 CalcMaxMana();
+	virtual int16 CalcMaxMana();
 	void SetGrid(int32 grid_){ grid=grid_; }
 	void SetSp2(uint32 sg2){ spawn_group=sg2; }
 	void SetWaypointMax(uint16 wp_){ wp_m=wp_; }
